fix: Consolidate ESLint and Pylint configs files
Refactored the ESLint and Pylint configuration files to improve
readability and maintainability. Removed superfluous comments and
whitelines from the Pylint configuration, providing a cleaner and more
concise setup. Added descriptive header to the ESLint configuration to
clarify its purpose and contents. These changes should make it easier
for new contributors to understand and work with the linting
configurations.